requirements.txt
anyio==4.4.0
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==42.0.0
debugpy==1.8.8
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.6
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
maritalk==0.2.6
matplotlib-inline==0.1.7
msal==1.26.0
msal-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.0.1
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
platformdirs==4.3.6
portalocker==2.8.2
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.31.0
safetensors==0.4.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.1.0
wcwidth==0.2.13
